July18update
-multythreading added
-not-defined nucleotided proceeding improved

+++ libs/WArray.h
@@ -4,7 +4,7 @@ using namespace std;
 
 struct WArray
 {
-	unsigned long long int value[4];
+	int value[4];
 	bool operator <(WArray &b)
 	{
 		return tie(this->value[0], this->value[1], this->value[2], this->value[3]) < tie(b.value[0], b.value[1], b.value[2], b.value[3]);
